requirements.txt
openai
requests
python-dotenv
dateparser
typer
